// frontend/src/types/movie.ts
export interface Movie {
    id: number;
    title: string;
    name?: string;
    overview: string;
    poster_path: string | null;
    backdrop_path: string | null;
    release_date: string;
    first_air_date?: string;
    vote_average: number;
}

export interface SearchResponse {
    page: number;
    results: Movie[];
    total_pages: number;
    total_results: number;
}